refactor(net): move `TracerChannel` to `channel.rs` module
