+++ src/tracing/net/ipv4.rs
@@ -1,6 +1,7 @@
 use crate::tracing::error::{TraceResult, TracerError};
+use crate::tracing::net::channel::MAX_PACKET_SIZE;
 use crate::tracing::net::platform::Ipv4TotalLengthByteOrder;
-use crate::tracing::net::{ProbeResponse, ProbeResponseData, MAX_PACKET_SIZE};
+use crate::tracing::net::{ProbeResponse, ProbeResponseData};
 use crate::tracing::packet::checksum::{icmp_ipv4_checksum, udp_ipv4_checksum};
 use crate::tracing::packet::icmpv4::destination_unreachable::DestinationUnreachablePacket;
 use crate::tracing::packet::icmpv4::echo_reply::EchoReplyPacket;
